common usage exceeded component
